refactor: Refactor Login to better match the composition API implementation
